Reconcile cross-domain core substrates
