+++ docs/data/multiphysics_capabilities.json
@@ -113,7 +113,14 @@
     "shared_particle_octree": "implemented",
     "isolated_barnes_hut_gravity": "implemented",
     "uniform_cartesian_fmm": "implemented",
-    "single_device_treepm_composition": "implemented"
+    "single_device_treepm_composition": "implemented",
+    "shared_dimensional_scale_contract": "implemented",
+    "shared_artifact_and_differentiation_contracts": "implemented",
+    "shared_observation_covariance_likelihood": "implemented",
+    "shared_direct_bh_fmm_ewald_treepm_gravity": "implemented",
+    "shared_coefficient_kdk_transaction": "implemented",
+    "shared_two_level_amr_mechanics": "implemented",
+    "shared_deterministic_event_replay": "implemented"
   },
   "inference": {
     "whitened_field_log_density": "implemented",
