+++ phydrax/_physical.py
@@ -1,0 +1,123 @@
+#
+# Copyright © 2026 PHYDRA, Inc. All rights reserved.
+#
+
+from __future__ import annotations
+
+from typing import Literal
+
+import equinox as eqx
+import numpy as np
+
+from ._fingerprint import canonical_fingerprint
+from ._strict import StrictModule
+from ._trainable import NonTrainableState
+
+
+LengthCoordinateKind = Literal["physical", "comoving", "code"]
+
+
+class DimensionalScaleContract(StrictModule, NonTrainableState):
+    """Shared length, mass, time, and coordinate-kind identity."""
+
+    length_unit: str = eqx.field(static=True)
+    mass_unit: str = eqx.field(static=True)
+    time_unit: str = eqx.field(static=True)
+    length_coordinate_kind: LengthCoordinateKind = eqx.field(static=True)
+    length_to_reference: float = eqx.field(static=True)
+    mass_to_reference: float = eqx.field(static=True)
+    time_to_reference: float = eqx.field(static=True)
+    scale_id: str = eqx.field(static=True)
+
+    def __init__(
+        self,
+        length_unit: str,
+        mass_unit: str,
+        time_unit: str,
+        /,
+        *,
+        length_coordinate_kind: LengthCoordinateKind = "physical",
+        length_to_reference: float = 1.0,
+        mass_to_reference: float = 1.0,
+        time_to_reference: float = 1.0,
+    ):
+        length = str(length_unit).strip()
+        mass = str(mass_unit).strip()
+        time = str(time_unit).strip()
+        kind = str(length_coordinate_kind).strip()
+        factors = tuple(
+            float(value)
+            for value in (
+                length_to_reference,
+                mass_to_reference,
+                time_to_reference,
+            )
+        )
+        if (
+            not length
+            or not mass
+            or not time
+            or kind not in ("physical", "comoving", "code")
+            or any(not np.isfinite(value) or value <= 0.0 for value in factors)
+        ):
+            raise ValueError("Dimensional scale contract is invalid.")
+        self.length_unit = length
+        self.mass_unit = mass
+        self.time_unit = time
+        self.length_coordinate_kind = kind
+        self.length_to_reference, self.mass_to_reference, self.time_to_reference = factors
+        self.scale_id = canonical_fingerprint(
+            {
+                "kind": "dimensional-scale-contract",
+                "length_unit": length,
+                "mass_unit": mass,
+                "time_unit": time,
+                "length_coordinate_kind": kind,
+                "length_to_reference": factors[0],
+                "mass_to_reference": factors[1],
+                "time_to_reference": factors[2],
+            }
+        )
+
+    @classmethod
+    def si(cls) -> DimensionalScaleContract:
+        return cls("m", "kg", "s", length_coordinate_kind="physical")
+
+    @property
+    def velocity_unit(self) -> str:
+        return f"{self.length_unit}/{self.time_unit}"
+
+    @property
+    def acceleration_unit(self) -> str:
+        return f"{self.length_unit}/{self.time_unit}^2"
+
+    @property
+    def gravitational_parameter_unit(self) -> str:
+        return f"{self.length_unit}^3/{self.time_unit}^2"
+
+    @property
+    def gravitational_constant_unit(self) -> str:
+        return f"{self.length_unit}^3/({self.mass_unit}*{self.time_unit}^2)"
+
+    @property
+    def hubble_unit(self) -> str:
+        return f"1/{self.time_unit}"
+
+    @property
+    def wavenumber_unit(self) -> str:
+        return f"1/{self.length_unit}"
+
+    @property
+    def power_spectrum_unit(self) -> str:
+        return f"{self.length_unit}^3"
+
+    @property
+    def potential_unit(self) -> str:
+        return f"{self.length_unit}^2/{self.time_unit}^2"
+
+    @property
+    def canonical_momentum_unit(self) -> str:
+        return f"{self.mass_unit}*{self.length_unit}/{self.time_unit}"
+
+
+__all__ = ["DimensionalScaleContract", "LengthCoordinateKind"]
