+++ phydrax/applications/astrodynamics/__init__.py
@@ -151,9 +151,7 @@ from ._scalable_gravity import (
     BarnesHutGravityPlan3D,
     CloseEncounterPolicy,
     detect_close_encounter,
-    DistributedTreePMPlan,
     EncounterEvaluation,
-    FastMultipoleGravityPlan3D,
     HierarchicalGravityResult,
     PreparedOctree3D,
 )
@@ -324,7 +322,6 @@ __all__ = [
     "ConjunctionPlan",
     "ConjunctionResult",
     "CoupledVehiclePlan",
-    "DistributedTreePMPlan",
     "DsstPlan",
     "DsstResult",
     "EarthOrientationEvaluation",
@@ -332,7 +329,6 @@ __all__ = [
     "EclipseGeometry",
     "EncounterEvaluation",
     "ExponentialAtmosphere",
-    "FastMultipoleGravityPlan3D",
     "FiniteBurnSegment",
     "FrameTransformEdge",
     "FrameTransformGraph",
