+++ phydrax/discretization/__init__.py
@@ -182,6 +182,7 @@ from .amr import (
     BlockLevelPlan,
     BlockLevelState,
     BlockMetadata,
+    CoarseFineFluxRegister,
     ConservativeAMRSubcyclingPlan,
     ConservativeBlockTransfer,
     FDAMRHaloPlan,
@@ -193,6 +194,8 @@ from .amr import (
     FluxRegister,
     PreparedFDAMRHierarchy,
     RefinementDecision,
+    TwoLevelAMRPlan,
+    TwoLevelAMRState,
 )
 from .collocation import chebyshev_lobatto_matrices, ChebyshevCollocation
 from .contact import (
@@ -1591,6 +1594,9 @@ __all__ = [
     "BlockLevelPlan",
     "BlockLevelState",
     "BlockMetadata",
+    "CoarseFineFluxRegister",
+    "TwoLevelAMRPlan",
+    "TwoLevelAMRState",
     "ClosureStencilKernel",
     "BlockInterface",
     "BlockSide",
